Solve errors and send a copy email to the admins
Solve problems with create and confirm meetings.
Solve problems with create and publish a project.
Solve the problem that not do the commit operation at the moment to
delete the project.
Send a copy of all the emails sending to the users to the admins of the project.
Change the user that receive the first email of the patent process.
Create a trigger which send a email to the users that the project was
created.
Solve errors to delete files projects.